refactor(app): remove external result question facade
