+++ packages/app/src/pages/layout/sidebar-items.tsx
@@ -6,6 +6,7 @@ import { Tooltip } from "@opencode-ai/ui/tooltip"
 import { A, useParams } from "@solidjs/router"
 import { type Accessor, createMemo, For, type JSX, Show } from "solid-js"
 import { formatKeybind } from "@/context/command"
+import { findDescendantExternalResultQuestion } from "@/context/global-sync/external-result-question"
 import { useGlobalSync } from "@/context/global-sync"
 import { useLanguage } from "@/context/language"
 import { useNotification } from "@/context/notification"
@@ -13,7 +14,6 @@ import { usePermission } from "@/context/permission"
 import { messageAgentColor } from "@/utils/agent"
 import { sessionTitle } from "@/utils/session-title"
 import { sessionPermissionRequest } from "../session/blockers/request-tree"
-import { findDescendantExternalResultQuestion } from "../session/blockers/running-external-result-question"
 import { createSessionRunning } from "../session/session-running-state"
 import { childSessionOnPath } from "./helpers"
 import { sidebarStatusKind } from "./sidebar-status-kind"
