--- packages/app/src/pages/session/blockers/running-external-result-question.ts
@@ -1,6 +1,0 @@
-export {
-  findDescendantExternalResultQuestion,
-  findRunningExternalResultQuestion,
-  rootSessionIDsWithDescendantExternalResultQuestions,
-} from "@/context/global-sync/external-result-question"
-export type { DockQuestionRequest, QuestionInfo } from "@/context/global-sync/external-result-question"
